+++ react/components/EditUserModal.tsx
@@ -1,11 +1,12 @@
 import React, { useState, useEffect } from 'react'
 import type { FunctionComponent } from 'react'
 import { Modal, Input, Button, Dropdown } from 'vtex.styleguide'
-import { defineMessages, useIntl, FormattedMessage } from 'react-intl'
+import { defineMessages, useIntl } from 'react-intl'
 import { useQuery } from 'react-apollo'
 
 import GET_ROLES from '../graphql/getRoles.graphql'
 import GET_COST_CENTERS from '../graphql/getCostCentersByOrganizationIdStorefront.graphql'
+import GET_COST_CENTERS_ADMIN from '../graphql/getCostCentersByOrganizationId.graphql'
 
 interface Props {
   loading: boolean
@@ -15,6 +16,7 @@ interface Props {
   handleRemoveUser: () => void
   handleCloseModal: () => void
   organizationId: string
+  isAdmin?: boolean
 }
 
 interface DropdownOption {
@@ -23,8 +25,12 @@ interface DropdownOption {
 }
 
 const storePrefix = 'store/b2b-organizations.'
+const adminPrefix = 'admin/b2b-organizations.'
 
-const messages = defineMessages({
+const storeMessages = defineMessages({
+  editUser: {
+    id: `${storePrefix}organization-details.edit-user`,
+  },
   save: {
     id: `${storePrefix}organization-details.button.save`,
   },
@@ -43,6 +49,42 @@ const messages = defineMessages({
   role: {
     id: `${storePrefix}user-details.placeholder-role`,
   },
+  userCostCenter: {
+    id: `${storePrefix}user-details.costCenter`,
+  },
+  userRole: {
+    id: `${storePrefix}user-details.role`,
+  },
+})
+
+const adminMessages = defineMessages({
+  editUser: {
+    id: `${adminPrefix}organization-details.edit-user`,
+  },
+  save: {
+    id: `${adminPrefix}organization-details.button.save`,
+  },
+  cancel: {
+    id: `${adminPrefix}organization-details.button.cancel`,
+  },
+  remove: {
+    id: `${adminPrefix}organization-details.button.remove-user`,
+  },
+  email: {
+    id: `${adminPrefix}user-details.email`,
+  },
+  costCenter: {
+    id: `${adminPrefix}user-details.placeholder-costCenter`,
+  },
+  role: {
+    id: `${adminPrefix}user-details.placeholder-role`,
+  },
+  userCostCenter: {
+    id: `${adminPrefix}user-details.costCenter`,
+  },
+  userRole: {
+    id: `${adminPrefix}user-details.role`,
+  },
 })
 
 const EditUserModal: FunctionComponent<Props> = ({
@@ -53,6 +95,7 @@ const EditUserModal: FunctionComponent<Props> = ({
   handleRemoveUser,
   handleCloseModal,
   organizationId,
+  isAdmin = false,
 }) => {
   const { formatMessage } = useIntl()
   const [userState, setUserState] = useState({} as UserDetails)
@@ -66,24 +109,31 @@ const EditUserModal: FunctionComponent<Props> = ({
     ssr: false,
   })
 
-  const { data: costCentersData } = useQuery(GET_COST_CENTERS, {
-    variables: { id: organizationId, pageSize: 100 },
-    fetchPolicy: 'network-only',
-    ssr: false,
-  })
+  const { data: costCentersData } = useQuery(
+    isAdmin ? GET_COST_CENTERS_ADMIN : GET_COST_CENTERS,
+    {
+      variables: { id: organizationId, pageSize: 100 },
+      fetchPolicy: 'network-only',
+      ssr: false,
+    }
+  )
 
   useEffect(() => {
     if (
-      !costCentersData?.getCostCentersByOrganizationIdStorefront?.data?.length
+      !costCentersData?.getCostCentersByOrganizationIdStorefront?.data
+        ?.length &&
+      !costCentersData?.getCostCentersByOrganizationId?.data?.length
     ) {
       return
     }
 
-    const options = costCentersData.getCostCentersByOrganizationIdStorefront.data.map(
-      (costCenter: any) => {
-        return { label: costCenter.name, value: costCenter.id }
-      }
-    )
+    const data = isAdmin
+      ? costCentersData.getCostCentersByOrganizationId.data
+      : costCentersData.getCostCentersByOrganizationIdStorefront.data
+
+    const options = data.map((costCenter: any) => {
+      return { label: costCenter.name, value: costCenter.id }
+    })
 
     setCostCenterOptions([...options])
   }, [costCentersData])
@@ -123,7 +173,9 @@ const EditUserModal: FunctionComponent<Props> = ({
               onClick={() => handleCloseModal()}
               disabled={loading}
             >
-              {formatMessage(messages.cancel)}
+              {formatMessage(
+                isAdmin ? adminMessages.cancel : storeMessages.cancel
+              )}
             </Button>
           </span>
           <span className="mr4">
@@ -132,7 +184,9 @@ const EditUserModal: FunctionComponent<Props> = ({
               onClick={() => handleRemoveUser()}
               disabled={loading}
             >
-              {formatMessage(messages.remove)}
+              {formatMessage(
+                isAdmin ? adminMessages.remove : storeMessages.remove
+              )}
             </Button>
           </span>
           <span>
@@ -144,7 +198,7 @@ const EditUserModal: FunctionComponent<Props> = ({
                 !userState.orgId || !userState.costId || !userState.roleId
               }
             >
-              {formatMessage(messages.save)}
+              {formatMessage(isAdmin ? adminMessages.save : storeMessages.save)}
             </Button>
           </span>
         </div>
@@ -154,12 +208,16 @@ const EditUserModal: FunctionComponent<Props> = ({
       closeOnOverlayClick={false}
     >
       <p className="f3 f1-ns fw3 gray">
-        <FormattedMessage id="store/b2b-organizations.organization-details.edit-user" />
+        {formatMessage(
+          isAdmin ? adminMessages.editUser : storeMessages.editUser
+        )}
       </p>
       <div className="w-100 mv6">
         <Input
           size="large"
-          label={formatMessage(messages.email)}
+          label={formatMessage(
+            isAdmin ? adminMessages.email : storeMessages.email
+          )}
           value={userState.email}
           disabled
         />
@@ -168,10 +226,16 @@ const EditUserModal: FunctionComponent<Props> = ({
         <Dropdown
           label={
             <h4 className="t-heading-5 mb0 pt3">
-              <FormattedMessage id="store/b2b-organizations.user-details.costCenter" />
+              {formatMessage(
+                isAdmin
+                  ? adminMessages.userCostCenter
+                  : storeMessages.userCostCenter
+              )}
             </h4>
           }
-          placeholder={formatMessage(messages.costCenter)}
+          placeholder={formatMessage(
+            isAdmin ? adminMessages.costCenter : storeMessages.costCenter
+          )}
           options={costCenterOptions}
           value={userState.costId}
           onChange={(_: any, v: string) =>
@@ -183,10 +247,14 @@ const EditUserModal: FunctionComponent<Props> = ({
         <Dropdown
           label={
             <h4 className="t-heading-5 mb0 pt3">
-              <FormattedMessage id="store/b2b-organizations.user-details.role" />
+              {formatMessage(
+                isAdmin ? adminMessages.userRole : storeMessages.userRole
+              )}
             </h4>
           }
-          placeholder={formatMessage(messages.role)}
+          placeholder={formatMessage(
+            isAdmin ? adminMessages.role : storeMessages.role
+          )}
           options={roleOptions}
           value={userState.roleId}
           onChange={(_: any, v: string) =>